requirements-test.txt
# Testing Framework
pytest==8.0.0
pytest-django==4.8.0
pytest-cov==5.0.0
pytest-xdist==3.5.0  # Parallel test execution
pytest-html==4.1.1   # HTML reports
pytest-json-report==1.5.0  # JSON output for visualization
factory-boy==3.3.0   # Test data factories
faker==24.0.0        # Realistic test data
coverage==7.4.0      # Code coverage
django-coverage-plugin==3.1.0
pytest-mock==3.12.0  # Mocking capabilities
freezegun==1.4.0     # Time mocking
responses==0.25.0    # HTTP mocking

# Visualization and Reporting
plotly==5.18.0       # Interactive charts
jinja2==3.1.3        # HTML templating
kaleido==0.2.1       # Static image export
pandas==2.3.0        # Visualization